# Add support for sccache.