@Firstyear 2021-05-14
> Shouldn't we also error out in case of the "--pkg-ccache" option?

Yep! Good catch, I have updated this.

> Minor: no need ot specify default=None.

Updated

> Hmm in case of an unset option, we should use the empty string '' instead of None. Otherwise, LGTM!

And also done! Thank you for the review! I have squashed and rebased to master as well :)